@nblumhardt 2016-05-31
Benchmarking `ForContext()`:

 * Baseline (measured in https://github.com/serilog/serilog/pull/750) is 1.08 s
 * This PR with the feature unused: 1.10 s (1.094 s, but rounded up as per the earlier numbers)
 * This PR with one override applied: 1.10 s (1.095 s, as above)

So, negligible difference in performance as far as I can tell. Fluctuation between runs is ~0.5 s :-)